requirements.txt
einops==0.3.2
ftfy==6.0.3
matplotlib==3.5.1
numpy==1.21.2
omegaconf==2.1.1
pickle5==0.0.12
Pillow==9.2.0
pytorch_lightning==1.5.7
regex>2022.1.18
scipy==1.7.3
torch==1.10.0
torchmetrics
torchvision
torch-fidelity==0.3.0
tqdm==4.61.2
scikit_learn>1.0.2
tokenizers>0.11.0
git+https://github.com/openai/CLIP.git